// orgado-client/src/interFace/interFace.ts
interface DropdownItem {
  link: string;
  title: string;
}

export interface NavMenuItem {
  id: number;
  link: string;
  title: string;
  hasDropdown?: boolean;
  megamenu?: boolean;
  dropdownItems?: DropdownItem[];
}